fix(security): raise ConversionError on script removal failure
CRITICAL: If script/style removal fails, returning original HTML could
leak API keys, secrets, or tracking code into markdown output.

- Add ConversionError exception type
- _remove_scripts_and_styles() raises ConversionError instead of
  returning potentially dangerous original HTML
- Add comprehensive tests for script/style removal

Also includes type fixes required for mypy to pass:
- float type for max_memory_mb (allows sub-MB limits in tests)
- Literal types for wait_for and backend in test fixtures
- type: ignore for MagicMock attribute setting

+++ src/sus/pipeline.py
@@ -56,7 +56,7 @@ class MemoryAwareQueue[T]:
         >>> result = await queue.get()
     """
 
-    def __init__(self, maxsize: int = 0, max_memory_mb: int = 500) -> None:
+    def __init__(self, maxsize: int = 0, max_memory_mb: float = 500) -> None:
         """Initialize memory-aware queue.
 
         Args:
